Added PadMaskedSequence, test
