+++ src/pydrobert/torch/_feats.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from turtle import forward
 from typing import Optional
 
 try:
