+++ src/pydrobert/torch/modules.py
@@ -48,6 +48,7 @@ __all__ = [
     "MixableSequentialLanguageModel",
     "MultiHeadedAttention",
     "OptimalCompletion",
+    "PadMaskedSequence",
     "PadVariable",
     "PolyharmonicSpline",
     "PrefixEditDistances",
@@ -82,7 +83,6 @@ from ._img import (
     PolyharmonicSpline,
     Warp1DGrid,
     SparseImageWarp,
-    PadVariable,
     RandomShift,
     SpecAugment,
 )
@@ -96,6 +96,7 @@ from ._mc import (
     LogisticBernoulliRebarControlVariate,
     GumbelOneHotCategoricalRebarControlVariate,
 )
+from ._pad import PadVariable, PadMaskedSequence
 from ._rl import TimeDistributedReturn
 from ._string import (
     EditDistance,
